create test config, niceify config loading & command structure
This doesn't really improve functionality a ton, but niceifies a couple
of things.

Configs are now loaded in a relatively sane way, don't require any more
from callers than the one function call, and loading them now does basic
validation.

The list command should serve as a usable template for creating further
commands, and the meat of it should be pretty testable. I'm debating
whether every command should have its "take the required arguments and
do the work" function have the same name - this is not easy to enforce
in Go, and really the only thing it helps is predictability when someone
new is looking at the implementation of a bunch of commands.

The test config files added under `dev-files` are super-basic to let me
do some basic poking of the commands.

Actual tests verifying stuff will be coming soon. Config loading might
not warrant a ton of focused & automated tests (for now) since setting
up enough valid config files to properly exercise whatever I end up
putting in there is probably not worth the effort.